requirements.txt
google-genai
dotenv
fastapi 
uvicorn